feat: implement quiz functionaly
Implement quiz component that prompts the user with questions, has a
button to show the answer, and then buttons to indicate whether they
answered correctly or not. After all cards in the deck are exhausted,
display their score and prompt them to either start the quiz again or
return to the home screen.